refactor / split code into separate files
